FOIA-280: Move selectedAgencies to form store.
