+++ js/stores/agency_component.js
@@ -23,7 +23,6 @@ class AgencyComponentStore extends Store {
       agencyComponents: new List(),
       agencyFinderDataComplete: false,
       agencyFinderDataProgress: 0,
-      selectedAgencies: [{ id: 0 }],
     };
   }
 
@@ -131,46 +130,6 @@ class AgencyComponentStore extends Store {
                 payload.agencyComponent,
               )),
             )),
-        });
-        this.__emitChange();
-        break;
-      }
-
-      case types.SELECTED_AGENCIES_UPDATE: {
-        const { selectedAgency, previousAgency } = payload;
-        const previousIsValid = typeof previousAgency === 'object'
-          && Object.prototype.hasOwnProperty.call(previousAgency, 'id');
-        const selectedIsValid = typeof selectedAgency === 'object'
-          && Object.prototype.hasOwnProperty.call(selectedAgency, 'id');
-
-        if (!selectedIsValid || !previousIsValid) {
-          break;
-        }
-
-        if (selectedAgency.id === previousAgency.id) {
-          break;
-        }
-
-        // Get a copy of the selected agencies state so that we don't
-        // mutate state.  Directly mutating the state prevents
-        // a rerender from firing, which causes multiple problems.
-        const selectedAgencies = [...this.state.selectedAgencies];
-        const indexToUpdate = selectedAgencies.findIndex(agency => agency.id === previousAgency.id);
-        selectedAgencies.splice(indexToUpdate, 1, selectedAgency);
-
-        Object.assign(this.state, {
-          selectedAgencies,
-        });
-        this.__emitChange();
-        break;
-      }
-
-      case types.SELECTED_AGENCIES_APPEND_BLANK: {
-        const selectedAgencies = [...this.state.selectedAgencies];
-        selectedAgencies.push({ id: (selectedAgencies.length) });
-
-        Object.assign(this.state, {
-          selectedAgencies,
         });
         this.__emitChange();
         break;
